% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{initialize_model}
\alias{initialize_model}
\title{Initialize a jags model prior to MCMC sampling.}
\usage{
initialize_model(m, n.chains = 1, n.adapt = 1000, quiet = FALSE)
}
\arguments{
\item{m}{A shollmcmcObj created by initialize_model()}

\item{n.chains}{The number of chains used for MCMC sampling. For details see ?rjags::jags.model.}

\item{n.adapt}{The number of adaptation iterations. For details see ?rjags::jags.model.}

\item{quiet}{For details see ?rjags::jags.model.}
}
\value{
A shollmcmcObj
}
\description{
Initialize a jags model prior to MCMC sampling.
}
\examples{
#See vignette("ungrouped_mouse_example", package = "ShollBayes") for examples.
}
