% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{initialize_model_ug3}
\alias{initialize_model_ug3}
\title{Fits the UG3 model}
\usage{
initialize_model_ug3(m, n.chains, n.adapt, quiet)
}
\arguments{
\item{m}{A shollmcmcObj created by initialize_model()}

\item{n.chains}{The number of chains used for MCMC sampling. For details see ?rjags::jags.model.}

\item{n.adapt}{The number of adaptation iterations. For details see ?rjags::jags.model.}

\item{quiet}{For details see ?rjags::jags.model.}
}
\description{
Fits the UG3 model
}
\keyword{internal}
