% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Load a pre-specified hierachical model for Sholl inference}
\usage{
load_model(model)
}
\arguments{
\item{model}{A string naming the desied model. Currenly supported models are "ug3_model", "mdnd_model", "crko_model", which correspond to the 3 examples in https://www.biorxiv.org/content/10.1101/2023.01.23.525256v1, respectively.}
}
\value{
A shollmcmcObj
}
\description{
Loads a pre-specified JAGS model
}
\examples{

m = load_model("ug3_model")

}
