% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_data.R
\name{restructure_data}
\alias{restructure_data}
\title{Restructures a data.frame into a format readable by rjags}
\usage{
restructure_data(m, ...)
}
\arguments{
\item{m}{A shollmcmcObj created by initialize_model()}

\item{...}{Additional arguments, see Details.}
}
\value{
A shollmcmcObj
}
\description{
Restructures a data.frame into a format readable by rjags
}
\details{
The required arguments for the currently supported models are included below:
test
another test
yet another test
}
\examples{
#See vignette("ungrouped_mouse_example", package = "ShollBayes") for examples.
}
