% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_data.R
\name{restructure_data_crko}
\alias{restructure_data_crko}
\title{Restructures data for fitting the CRKO model}
\usage{
restructure_data_crko(y, R, Cell_ID, Group_ID, Animal_ID, Genotype_ID)
}
\arguments{
\item{y}{Vector of intersections}

\item{R}{Vector of radii}

\item{Cell_ID}{ID vector indexing cells}

\item{Group_ID}{ID vector indexing treatment group}

\item{Animal_ID}{ID vector indexing animals}

\item{Genotype_ID}{ID vector indexing genotype}
}
\description{
Restructures data for fitting the CRKO model
}
\keyword{internal}
