% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_data.R
\name{restructure_data_mdnd}
\alias{restructure_data_mdnd}
\title{Restructures data for fitting the MDND model}
\usage{
restructure_data_mdnd(y, R, Animal_ID, Group_ID)
}
\arguments{
\item{y}{Vector of intersections}

\item{R}{Vector of radii}

\item{Animal_ID}{ID vector indexing animals}

\item{Group_ID}{ID vector indexing treatment group}
}
\description{
Restructures data for fitting the MDND model
}
\keyword{internal}
