% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_data.R
\name{restructure_data_ug3}
\alias{restructure_data_ug3}
\title{Restructures data for fitting the UG3 model}
\usage{
restructure_data_ug3(y, R, Cell_ID, Image_ID, Animal_ID)
}
\arguments{
\item{y}{Vector of intersections}

\item{R}{Vector of radii}

\item{Cell_ID}{ID vector indexing cells}

\item{Image_ID}{ID vector indexing images}

\item{Animal_ID}{ID vector indexing animals}
}
\description{
Restructures data for fitting the UG3 model
}
\keyword{internal}
