% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shollMCMC_class.R
\name{sample<-,shollmcmcObj-method}
\alias{sample<-,shollmcmcObj-method}
\title{sample<-}
\usage{
\S4method{sample}{shollmcmcObj}(x) <- value
}
\arguments{
\item{x}{An object of class shollmcmcObj}

\item{value}{Value to assign to slot.}
}
\description{
Set the sample slot in an object of class shollmcmcObj
}
\examples{

#See vignette("ungrouped_mouse_example", package = "ShollBayes") for examples.

}
