% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shollMCMC_class.R
\name{sample,shollmcmcObj-method}
\alias{sample,shollmcmcObj-method}
\title{stack()}
\usage{
\S4method{sample}{shollmcmcObj}(x)
}
\arguments{
\item{x}{An object of class shollmcmcObj}
}
\value{
The sample slot of shollmcmcObj
}
\description{
Retrieve the sample slot from an object of class shollmcmcObj
}
\examples{

#See vignette("ungrouped_mouse_example", package = "ShollBayes") for examples.

}
