% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{sample_model}
\alias{sample_model}
\title{Obtain MCMC samples from a jags model}
\usage{
sample_model(
  m,
  variable.name = "All",
  n.iter,
  thin = 1,
  na.rm = FALSE,
  new_sample = TRUE,
  ...
)
}
\arguments{
\item{m}{A shollmcmcObj created by initialize_model()}

\item{variable.name}{A string vector specifying which parameters should be monitored during sampling. Default is "All", which monitors all parameters. We recommend using this argument unless computationally limited}

\item{n.iter}{Number of MCMC samples to obtain.}

\item{thin}{Thinning interval}

\item{na.rm}{Logical flag that indicates whether variables containing missing values should be omitted. See ?coda::coda.samples for details.}

\item{new_sample}{Logical. If TRUE any existing samples in the samples slot will be overwritten, if FALSE they will be appended.}

\item{...}{Additional arguments. See rjags::coda.samples.}
}
\value{
A shollmcmcObj
}
\description{
Obtain MCMC samples from a jags model
}
\examples{
#See vignette("ungrouped_mouse_example", package = "ShollBayes") for examples.
}
