% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shollMCMC_class.R
\docType{class}
\name{shollmcmcObj-class}
\alias{shollmcmcObj-class}
\title{An S4 object for storing data and models for a Sholl analysis dataset.}
\description{
An S4 object for storing data and models for a Sholl analysis dataset.
}
\keyword{internal}
