% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{update_model}
\alias{update_model}
\title{A function to burn-in a specified number of iterations.}
\usage{
update_model(m, n.iter, ...)
}
\arguments{
\item{m}{A shollmcmcObj created by initialize_model()}

\item{n.iter}{Number of burn-in iterations}

\item{...}{Additional arguments. See rjags::update.}
}
\value{
A shollmcmcObj
}
\description{
A function to burn-in a specified number of iterations.
}
\examples{

#See vignette("ungrouped_mouse_example", package = "ShollBayes") for examples.


}
